switch the threshold lab to the sinter version
